Check that the passed timestamp value is remotely sane
The document is being signed now; so if someone sends data for a
document signature that is way too far from the current date and time,
reject that.

Allow for a 5 minute margin in both directions to account for clock
skew, however.

This reverts 89643ab9402fb0e2c91674b3adaa400ad69f9d7b