Make StoredKey serializeable with Jackson
We tried to deserialize a SecretKey, but you can't do that.
Additionally, the "isTooOld()" boolean method was being stored as a
property, which makes no sense.

Instead, hide the properties that shouldn't be stored with the
JacksonIgnore annotation, and add another property that only contains
the encoded secret key, which we can then later parse with SecretKeySpec
rather than trying (and failing) to deserialize a SecretKey.